Innershift/requirements.txt
# --- Traitement du Langage Naturel ---
spacy==3.5.3
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.5.0/fr_core_news_sm-3.5.0-py3-none-any.whl
nltk==3.8.1
transformers==4.32.1
sentence-transformers==2.2.2

# --- Machine Learning et Math ---
scikit-learn==1.2.2
numpy==1.23.5
pandas==1.5.3

# --- Traitement Audio ---
sounddevice==0.5.1
scipy==1.11.4
openai-whisper==20231117
torch==2.0.1
tiktoken==0.9.0
ffmpeg-python==0.2.0
pydub==0.25.1
speechrecognition==3.8.1

# --- Visualisation et Graphiques ---
matplotlib==3.7.1
seaborn==0.12.2
huggingface-hub==0.15.1

# --- Génération de PDF ---
fpdf==1.7.2

# --- Web (API future) ---
fastapi==0.95.1
uvicorn==0.22.0

# --- Optionnel (vision & émotions future) ---
opencv-python==4.7.0.72
mediapipe==0.10.21
